+++ lua/github-theme/group/modules/indent_blankline.lua
@@ -5,13 +5,16 @@ local C = require('github-theme.lib.color')
 local M = {}
 
 function M.get(spec, config, opts)
-  local fg = C(spec.bg1):blend(C(spec.fg1), 0.2):to_css()
-  local blue_fg = C(spec.bg1):blend(C(spec.palette.cyan.bright), 0.4):to_css()
+  local p = spec.palette
+
+  local fg = C(spec.bg1):blend(C(spec.fg1), 0.1):to_css()
+  local cyan_fg = C(spec.bg1):blend(C(p.cyan.bright), 0.35):to_css()
 
   -- stylua: ignore
   return {
     IndentBlanklineChar = { fg = fg  },
-    IndentBlanklineContextChar = { fg = blue_fg },
+    IndentBlanklineContextChar = { fg = cyan_fg },
+    IndentBlanklineContextStart = { style = 'underline', sp = p.orange },
   }
 end