+++ lua/github-theme/lib/compiler.lua
@@ -59,6 +59,10 @@ vim.o.background = "%s"
       op.bg = values.bg
       op.fg = values.fg
       op.sp = values.sp
+      op.blend = values.blend
+      if op.nocombine == nil then
+        op.nocombine = values.nocombine
+      end
       table.insert(lines, fmt([[h(0, "%s", %s)]], name, inspect(op)))
     end
   end