+++ lua/github-theme/palette/github_dark_dimmed.lua
@@ -5,49 +5,11 @@ local meta = {
   light = false,
 }
 
----Github Dark Dimmed scale variables
----source: https://github.com/primer/primitives/blob/main/data/colors/themes/dark_dimmed.ts
--- stylua: ignore
-local scale = {
-  black = '#1c2128',
-  white = '#cdd9e5',
-  gray = {
-    '#cdd9e5', '#adbac7', '#909dab', '#768390', '#636e7b',
-    '#545d68', '#444c56', '#373e47', '#2d333b', '#22272e',
-  },
-  blue = {
-    '#c6e6ff', '#96d0ff', '#6cb6ff', '#539bf5', '#4184e4',
-    '#316dca', '#255ab2', '#1b4b91', '#143d79', '#0f2d5c',
-  },
-  green = {
-    '#b4f1b4', '#8ddb8c', '#6bc46d', '#57ab5a', '#46954a',
-    '#347d39', '#2b6a30', '#245829', '#1b4721', '#113417',
-  },
-  yellow = {
-    '#fbe090', '#eac55f', '#daaa3f', '#c69026', '#ae7c14',
-    '#966600', '#805400', '#6c4400', '#593600', '#452700',
-  },
-  orange = {
-    '#ffddb0', '#ffbc6f', '#f69d50', '#e0823d', '#cc6b2c',
-    '#ae5622', '#94471b', '#7f3913', '#682d0f', '#4d210c',
-  },
-  red = {
-    '#ffd8d3', '#ffb8b0', '#ff938a', '#f47067', '#e5534b',
-    '#c93c37', '#ad2e2c', '#922323', '#78191b', '#5d0f12',
-  },
-  purple = {
-    '#eedcff', '#dcbdfb', '#dcbdfb', '#b083f0', '#986ee2',
-    '#8256d0', '#6b44bc', '#5936a2', '#472c82', '#352160',
-  },
-  pink = {
-    '#ffd7eb', '#ffb3d8', '#fc8dc7', '#e275ad', '#c96198',
-    '#ae4c82', '#983b6e', '#7e325a', '#69264a', '#551639',
-  },
-  coral = {
-    '#ffdacf', '#ffb9a5', '#f79981', '#ec775c', '#de5b41',
-    '#c2442d', '#a93524', '#8d291b', '#771d13', '#5d1008',
-  },
-}
+local primitives =
+  require('github-theme.palette.primitives.' .. meta.name:gsub('^github%W*', '', 1))
+
+local pl = primitives.prettylights
+local scale = primitives.scale
 
 C.WHITE = C(scale.white)
 C.BLACK = C(scale.black)
@@ -173,39 +135,39 @@ local function generate_spec(pal)
     bg3  = alpha(C(pal.fg.default), 0.1),            -- Lighter bg (cursor line)
     bg4  = pal.scale.gray[4],                        -- Conceal
 
-    fg0  = pal.fg.subtle,                            -- Lighter fg
-    fg1  = pal.fg.default,                           -- Default fg
-    fg2  = pal.fg.muted,                             -- Darker fg (status line)
-    fg3  = pal.scale.gray[5],                        -- Darker fg (line numbers, fold columns)
+    fg0  = pal.fg.subtle,                               -- Lighter fg
+    fg1  = pal.fg.default,                              -- Default fg
+    fg2  = pal.fg.muted,                                -- Darker fg (status line)
+    fg3  = pal.scale.gray[5],                           -- Darker fg (line numbers, fold columns)
 
-    sel0 = alpha(C(pal.accent.fg), 0.40),            -- Visual selection bg
-    sel1 = alpha(C(pal.accent.muted), 0.90),         -- Popup sel bg
-    sel2 = alpha(C(pal.scale.yellow[1]), 0.20),      -- Search bg
+    sel0 = alpha(C(pal.accent.fg), 0.40),               -- Visual selection bg
+    sel1 = alpha(C(pal.accent.muted), 0.90),            -- Popup sel bg
+    sel2 = alpha(C(pal.scale.yellow[1]), 0.20),         -- Search bg
   }
 
   spec.syntax = {
-    bracket     = pal.scale.green[4],                -- Brackets and Punctuation
-    builtin0    = pal.scale.red[4],                  -- Builtin variable (Return Keywords, Regex, etc.)
-    builtin1    = pal.scale.red[4],                  -- Builtin type
-    builtin2    = pal.scale.blue[3],                 -- Builtin const
-    comment     = pal.scale.gray[5],                 -- Comment
-    conditional = pal.scale.red[4],                  -- Conditional and loop
-    const       = pal.scale.blue[3],                 -- Constants, imports and booleans
-    dep         = pal.scale.red[3],                  -- Deprecated
-    field       = pal.scale.purple[3],               -- Field
-    func        = pal.scale.purple[2],               -- Functions and Titles
-    ident       = pal.scale.blue[3],                 -- Identifiers
-    keyword     = pal.scale.red[4],                  -- Keywords
-    number      = pal.scale.blue[3],                 -- Numbers
-    operator    = pal.scale.red[4],                  -- Operators
-    param       = pal.scale.orange[3],               -- Parameters
-    preproc     = pal.scale.red[4],                  -- PreProc
-    regex       = pal.scale.blue[3],                 -- Regex
-    statement   = pal.scale.red[4],                  -- Statements
-    string      = pal.scale.blue[2],                 -- Strings
-    type        = pal.scale.red[4],                  -- Types
-    tag         = pal.scale.green[2],                -- Tags
-    variable    = spec.fg1,                          -- Variables
+    bracket     = spec.fg1,                             -- Brackets and Punctuation
+    builtin0    = pl.syntax.constant,                   -- Builtin variable
+    builtin1    = pl.syntax.keyword,                    -- Builtin type
+    builtin2    = pl.syntax.constant,                   -- Builtin const
+    comment     = pl.syntax.comment,                    -- Comment
+    conditional = pl.syntax.keyword,                    -- Conditional and loop
+    const       = pl.syntax.constant,                   -- Constants, imports and booleans
+    dep         = pal.scale.red[3],                     -- Deprecated
+    field       = pl.syntax.constant,                   -- Field
+    func        = pl.syntax.entity,                     -- Functions and Titles
+    ident       = spec.fg1,                             -- Identifiers
+    keyword     = pl.syntax.keyword,                    -- Keywords
+    number      = pl.syntax.constant,                   -- Numbers
+    operator    = pl.syntax.constant,                   -- Operators
+    param       = spec.fg1,                             -- Parameters
+    preproc     = pl.syntax.keyword,                    -- PreProc
+    regex       = pl.syntax.string,                     -- Regex
+    statement   = pl.syntax.keyword,                    -- Statements
+    string      = pl.syntax.string,                     -- Strings
+    type        = pl.syntax.variable,                   -- Types
+    tag         = pl.syntax.entityTag,                  -- Tags
+    variable    = spec.fg1,                             -- Variables
   }
 
   spec.diag = {
@@ -241,4 +203,9 @@ local function generate_spec(pal)
   return spec
 end
 
-return { meta = meta, palette = palette, generate_spec = generate_spec }
+return {
+  meta = meta,
+  primitives = primitives,
+  palette = palette,
+  generate_spec = generate_spec,
+}