+++ lua/github-theme/palette/github_light_high_contrast.lua
@@ -5,49 +5,11 @@ local meta = {
   light = true,
 }
 
----Github Light High Contrast scale variables
----source: https://github.com/primer/primitives/blob/main/data/colors/themes/light_high_contrast.ts
--- stylua: ignore
-local scale = {
-  black = '#010409',
-  white = '#ffffff',
-  gray = {
-    '#ffffff', '#e7ecf0', '#ced5dc', '#acb6c0', '#88929d',
-    '#66707b', '#4b535d', '#343b43', '#20252c', '#0e1116',
-  },
-  blue = {
-    '#dff7ff', '#9cd7ff', '#67b3fd', '#368cf9', '#1168e3',
-    '#0349b4', '#023b95', '#022f7a', '#032563', '#021a4a',
-  },
-  green = {
-    '#d2fedb', '#82e596', '#43c663', '#26a148', '#117f32',
-    '#055d20', '#024c1a', '#013d14', '#003110', '#00230b',
-  },
-  yellow = {
-    '#fcf7be', '#f0ce53', '#d5a824', '#b58407', '#956400',
-    '#744500', '#603700', '#4e2c00', '#3f2200', '#2e1800',
-  },
-  orange = {
-    '#fff2d5', '#ffc67b', '#f99636', '#dc6d1a', '#b45105',
-    '#873800', '#702c00', '#5b2300', '#491b00', '#361200',
-  },
-  red = {
-    '#fff0ee', '#ffc1bc', '#ff8e8a', '#ee5a5d', '#d5232c',
-    '#a0111f', '#86061d', '#6e011a', '#5a0016', '#430011',
-  },
-  purple = {
-    '#faf0fe', '#e0c5ff', '#c49bff', '#a371f7', '#844ae7',
-    '#622cbc', '#512598', '#411d7b', '#341763', '#260f49',
-  },
-  pink = {
-    '#feeff7', '#ffbde0', '#fc87ca', '#ed4baf', '#c9248e',
-    '#971368', '#7d0c57', '#660847', '#53043a', '#3e022b',
-  },
-  coral = {
-    '#fff0ed', '#ffc2b6', '#ff8f7e', '#ef5b48', '#cd3425',
-    '#9f1710', '#870706', '#6f0107', '#5b0002', '#430200',
-  },
-}
+local primitives =
+  require('github-theme.palette.primitives.' .. meta.name:gsub('^github%W*', '', 1))
+
+local pl = primitives.prettylights
+local scale = primitives.scale
 
 C.WHITE = C(scale.white)
 C.BLACK = C(scale.black)
@@ -168,44 +130,44 @@ local palette = {
 local function generate_spec(pal)
   -- stylua: ignore start
   local spec = {
-    bg0  = pal.canvas.inset,                        -- Dark bg (popup and float)
-    bg1  = pal.canvas.default,                      -- Default bg
-    bg2  = alpha(C(pal.neutral.emphasis), 0.1),     -- Lighter bg (colorcolumn Folds)
-    bg3  = pal.scale.gray[2],                       -- Lighter bg (cursor line)
-    bg4  = pal.border.default,                      -- Conceal
+    bg0  = pal.canvas.inset,                            -- Dark bg (popup and float)
+    bg1  = pal.canvas.default,                          -- Default bg
+    bg2  = alpha(C(pal.neutral.emphasis), 0.1),         -- Lighter bg (colorcolumn Folds)
+    bg3  = pal.scale.gray[2],                           -- Lighter bg (cursor line)
+    bg4  = pal.border.default,                          -- Conceal
 
-    fg0  = pal.fg.subtle,                           -- Lighter fg
-    fg1  = pal.fg.default,                          -- Default fg
+    fg0  = pal.fg.subtle,                               -- Lighter fg
+    fg1  = pal.fg.default,                              -- Default fg
     fg2  = pal.scale.gray[9],                           -- Darker fg (status line)
-    fg3  = pal.scale.gray[7],                       -- Darker fg (line numbers, fold columns)
+    fg3  = pal.scale.gray[7],                           -- Darker fg (line numbers, fold columns)
 
-    sel0 = alpha(C(pal.accent.fg), 0.2),            -- Visual selection bg
-    sel1 = alpha(C(pal.accent.muted), 0.90),        -- Popup sel bg
-    sel2 = alpha(C(pal.attention.emphasis), 0.3)    -- Search bg
+    sel0 = alpha(C(pal.accent.fg), 0.2),                -- Visual selection bg
+    sel1 = alpha(C(pal.accent.muted), 0.90),            -- Popup sel bg
+    sel2 = alpha(C(pal.attention.emphasis), 0.3)        -- Search bg
   }
 
   spec.syntax = {
-    bracket     = pal.scale.orange[5],                  -- Brackets and Punctuation
-    builtin0    = pal.scale.red[6],                     -- Builtin variable
-    builtin1    = pal.scale.red[6],                     -- Builtin type
-    builtin2    = pal.scale.blue[7],                    -- Builtin const
-    comment     = pal.scale.gray[6],                    -- Comment
-    conditional = pal.scale.red[6],                     -- Conditional and loop
-    const       = pal.scale.blue[6],                    -- Constants, imports and booleans
+    bracket     = spec.fg1,                             -- Brackets and Punctuation
+    builtin0    = pl.syntax.constant,                   -- Builtin variable
+    builtin1    = pl.syntax.keyword,                    -- Builtin type
+    builtin2    = pl.syntax.constant,                   -- Builtin const
+    comment     = pl.syntax.comment,                    -- Comment
+    conditional = pl.syntax.keyword,                    -- Conditional and loop
+    const       = pl.syntax.constant,                   -- Constants, imports and booleans
     dep         = pal.scale.red[8],                     -- Deprecated
-    field       = spec.fg1,                             -- Field
-    func        = pal.scale.purple[6],                  -- Functions and Titles
-    ident       = pal.scale.blue[9],                    -- Identifiers
-    keyword     = pal.scale.red[6],                     -- Keywords
-    number      = pal.scale.blue[7],                    -- Numbers
-    operator    = pal.scale.red[6],                     -- Operators
-    param       = pal.scale.orange[5],                  -- PreProc
-    preproc     = pal.scale.red[6],                     -- PreProc
-    regex       = pal.scale.blue[9],                    -- Regex
-    statement   = pal.scale.red[6],                     -- Statements
-    string      = pal.scale.blue[8],                    -- Strings
-    type        = pal.scale.red[6],                     -- Types
-    tag         = pal.scale.green[6],                   -- Tags
+    field       = pl.syntax.constant,                   -- Field
+    func        = pl.syntax.entity,                     -- Functions and Titles
+    ident       = spec.fg1,                             -- Identifiers
+    keyword     = pl.syntax.keyword,                    -- Keywords
+    number      = pl.syntax.constant,                   -- Numbers
+    operator    = pl.syntax.constant,                   -- Operators
+    param       = spec.fg1,                             -- Parameters
+    preproc     = pl.syntax.keyword,                    -- PreProc
+    regex       = pl.syntax.string,                     -- Regex
+    statement   = pl.syntax.keyword,                    -- Statements
+    string      = pl.syntax.string,                     -- Strings
+    type        = pl.syntax.variable,                   -- Types
+    tag         = pl.syntax.entityTag,                  -- Tags
     variable    = spec.fg1,                             -- Variables
   }
 
@@ -242,4 +204,9 @@ local function generate_spec(pal)
   return spec
 end
 
-return { meta = meta, palette = palette, generate_spec = generate_spec }
+return {
+  meta = meta,
+  primitives = primitives,
+  palette = palette,
+  generate_spec = generate_spec,
+}