+++ queries/php/highlights.scm
@@ -1,0 +1,6 @@
+;; extends
+
+((php_tag) @tag)
+
+(variable_name
+  "$" @operator)